ref: moved disconnects_for_trivial_issues
