+++ src/tests/conformance/invalid_message/mod.rs
@@ -1,1 +1,2 @@
+mod disconnect;
 mod reject;
